accountant added

+++ database/migrations/2022_07_20_155508_create_billings_table.php
@@ -15,7 +15,7 @@ class CreateBillingsTable extends Migration
     {
         Schema::create('billings', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('patient_id');
+            $table->unsignedBigInteger('patient_id');
             $table->datetime('bill_date');
             $table->double('amount')->nullable();
             $table->timestamps();
